requirements.txt
mfrc522==0.0.7
RPi.GPIO==0.6.5
spidev==3.4
